server/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
google-genai
httpx
slowapi
